src/UpsCoolWeb.Components/Mvc/Attributes/AjaxOnlyAttribute.cs
using Microsoft.AspNetCore.Mvc.Abstractions;
using Microsoft.AspNetCore.Mvc.ActionConstraints;
using Microsoft.AspNetCore.Routing;
using System;

namespace UpsCoolWeb.Components.Mvc
{
    [AttributeUsage(AttributeTargets.Method)]
    public class AjaxOnlyAttribute : ActionMethodSelectorAttribute
    {
        public override Boolean IsValidForRequest(RouteContext context, ActionDescriptor action)
        {
            return context.HttpContext.Request.Headers["X-Requested-With"] == "XMLHttpRequest";
        }
    }
}
